Add requests

+++ src/main/java/com/example/real/estate/agency/repository/UserRoleRepository.java
@@ -3,6 +3,6 @@ package com.example.real.estate.agency.repository;
 import com.example.real.estate.agency.entity.UserRole;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRoleRepository extends JpaRepository<UserRole, Integer> {
+public interface UserRoleRepository extends JpaRepository<UserRole, Long> {
     UserRole getUserRoleById(Long userId);
 }
